// src/app/models/dateo.ts
export class Dateo {
  idDateo: number;
  address: string;
  description: string;
  latitude: string;
  longitude: string;
  tag1: string;
  tag2: string;
  tag3: string;
  photoUrl: string;
  creationDateTime: string;
  userId: number;
  userFirstName: string;
  userLastName: string;
  userEmail: string;

  constructor() {}

  /*constructor(public idDateo: number, public address: string, public description: string, public latitude: number,
              public longitude: number, public tag1: string, public tag2: string, public tag3: string, public photoUrl: string,
              public creationDateTime: string, userId: number, userName: string) { } */
}